rename new-proposal-review-view to proposal-review-view; old one becomes old-
