+++ src/app/[orgSlug]/study/[studyId]/review/page.tsx
@@ -7,7 +7,7 @@ import { getStudyAction } from '@/server/actions/study.actions'
 import { sessionFromClerk } from '@/server/clerk'
 import { redirect } from 'next/navigation'
 import { CodeReviewView } from './code-review-view'
-import { ProposalReviewView } from './proposal-review-view'
+import { OldProposalReviewView } from './old-proposal-review-view'
 
 export default async function StudyReviewPage(props: {
     params: Promise<{
@@ -41,7 +41,7 @@ export default async function StudyReviewPage(props: {
         // instead of the code review — they've already reviewed code and need to revisit the proposal
         if (searchParams.from === 'agreements' && codeSubmitted) {
             return (
-                <ProposalReviewView
+                <OldProposalReviewView
                     orgSlug={orgSlug}
                     study={study}
                     agreementsHref={Routes.studyAgreements({ orgSlug, studyId })}
@@ -55,7 +55,7 @@ export default async function StudyReviewPage(props: {
             }
             return <CodeReviewView orgSlug={orgSlug} study={study} />
         }
-        return <ProposalReviewView orgSlug={orgSlug} study={study} />
+        return <OldProposalReviewView orgSlug={orgSlug} study={study} />
     }
 
     return <AlertNotFound title="Study was not found" message="no such study exists" />
